+++ packages/shsp/lib/src/impl/connection/connection.dart
@@ -2,6 +2,5 @@
 
 library;
 
-export 'handshake_ip.dart';
 export 'handshake_ownership.dart';
 export 'handshake_time.dart';