+++ packages/shsp/lib/src/interfaces/socket/i_shsp_socket_migratable.dart
@@ -3,6 +3,6 @@
 import '../../../shsp.dart';
 
 /// Interface for SHSP Socket
-abstract interface class IShspSocketWrapper implements IShspSocket{
+abstract interface class IShspSocketMigratable implements IShspSocket{
   void migrateSocket(IShspSocket socket);
 }